+++ test/unit/pool/test-debug.js
@@ -1,6 +1,7 @@
 var assert = require('assert');
 var common = require('../../common');
 var pool   = common.createPool({debug: true, port: common.fakeServerPort});
+var util   = require('util');
 
 var tid    = 0;
 var server = common.createFakeServer();
@@ -10,9 +11,10 @@ server.listen(common.fakeServerPort, function (err) {
 
   var messages = [];
 
-  console.log = function (str) {
-    if (typeof str === 'string' && str.length !== 0) {
-      messages.push(str);
+  console.log = function () {
+    var msg = util.format.apply(this, arguments);
+    if (String(msg).indexOf('--') !== -1) {
+      messages.push(msg.split(' {')[0]);
     }
   };